재료 추가시 테스트 코드 fail 픽스
재료 추가시 Dto와 entity 사이의 기본형, 참조형 타입 차이로 인한 테스트 코드 fail 부분 픽스 했습니다.

+++ src/main/java/com/my/bob/core/domain/refrigerator/dto/response/RefrigeratorIngredientDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 public class RefrigeratorIngredientDto {
     // 조회용
 
-    private long ingredientId;   // refrigerator_ingredient_id
+    private Integer ingredientId;   // refrigerator_ingredient_id
 
     private String ingredientName;
 
@@ -18,7 +18,7 @@ public class RefrigeratorIngredientDto {
     private String addedDate;
 
     @Builder
-    public RefrigeratorIngredientDto(long ingredientId, String ingredientName, String ingredientUrl,
+    public RefrigeratorIngredientDto(Integer ingredientId, String ingredientName, String ingredientUrl,
                                      String addedDate) {
         this.ingredientId = ingredientId;
         this.ingredientName = ingredientName;
